# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "COMPANY-CMS",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/master_dig_tech/services": {
            "get": {
                "tags": [
                    "Master Digital Technologies"
                ],
                "summary": "Все услуги",
                "description": "Returns a collection of services filtered by project ID 3.",
                "operationId": "90514d5243dd44e265c7ab0fce4983cd",
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "No services found for the specified project",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/master_dig_tech/services/{service_slug}": {
            "get": {
                "tags": [
                    "Master Digital Technologies"
                ],
                "summary": "Услуга по slug",
                "description": "Returns a service resource for the specified slug.",
                "operationId": "19dd0db2112194e84c114dd18d37fc4c",
                "parameters": [
                    {
                        "name": "service_slug",
                        "in": "path",
                        "description": "Slug of the service",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Service not found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/services/{service_slug}/{category_slug}": {
            "get": {
                "tags": [
                    "Master Digital Technologies"
                ],
                "summary": "Получить контент по услугам и категориям",
                "description": "Возвращает фотографии и видео для заданной услуги и категории.",
                "operationId": "549095ad0b12ffe194d7aea5f65c15c5",
                "parameters": [
                    {
                        "name": "serviceSlug",
                        "in": "query",
                        "description": "Слаг услуги",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "categorySlug",
                        "in": "query",
                        "description": "Слаг категории",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "photos": {
                                            "type": "array",
                                            "items": {}
                                        },
                                        "videos": {
                                            "type": "array",
                                            "items": {}
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Услуга или категория не найдены"
                    }
                }
            }
        },
...
    },
...
}